+++ BuildMasterAutomation/Functions/Get-BMRaftTypeDisplayName.ps1
@@ -4,11 +4,17 @@ function Get-BMRaftTypeDisplayName
     [CmdletBinding()]
     param(
         [Parameter(Mandatory, ValueFromPipeline)]
+        [AllowNull() ]
         [BMRaftItemTypeCode] $TypeCode
     )
 
     process
     {
+        if ($null -eq $TypeCode)
+        {
+            return 'Raft Item'
+        }
+
         switch ($TypeCode)
         {
             'DeploymentPlan' { return 'Deployment Plan' }